fashion_dataset/requirements.txt
tensorflow
pandas
pillow
scikit-learn
streamlit
numpy
matplotlib
